pyproject.toml
[project]
name = "gpu-ttd-dr-refactored"
version = "0.1.0"
description = "Modularized TTD-DR GPU report agent (FastAPI + Streamlit)"
authors = [{name="OncleJura", email="hosing87@naver.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2",
    "python-dotenv",
    "requests",
    "beautifulsoup4",
    "trafilatura; platform_system != 'Windows'",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "chromadb",
]

[tool.setuptools]
package-dir = {"" = "."}
